Add doc tests to 10 components with lowest coverage
Add 69 new doc tests across button, chart, checkbox, data_grid,
file_browser, loading_list, log_viewer, pane_layout, scrollable_text,
and step_indicator components. Doc tests cover public accessor methods,
builder methods, state manipulation, event handling, and instance
methods.

Extract log_viewer rendering functions into view.rs submodule to keep
mod.rs under the 1000-line limit while using only the public API.
